requirements.txt
# Core ML/AI libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
safetensors>=0.3.0

# Data processing and utilities
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0

# Web scraping and API
praw>=7.7.0
requests>=2.31.0

# Web application
streamlit>=1.25.0

# Hugging Face integration
huggingface_hub>=0.16.0

# Environment management
python-dotenv>=1.0.0

# Optional: For better performance on GPU
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118 